automatically use .config/tacos-gha/slices.allowlist
